requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
frozenlist==1.3.3
multidict==6.0.4
openai==0.27.6
yarl==1.9.2
colorama==0.4.6
python-dotenv==1.0.0
google-api-python-client==2.86.0